backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "x_x"
authors = ["Тимофей Городилов <timotheos94@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.0"
flake8 = "^6.0.0"
mypy = "^0.991"
django = "4.1"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"
isort = "^5.11.4"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^4.3"
djoser = "^2.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
